PhotoNotes/photonotes/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.13.1
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.5
pymemcache==4.0.0
python-dotenv==0.21.0
python-memcached==1.59
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
